[Fabric] Clean up a bunch of include redirect files
