--- vnext/Microsoft.ReactNative/Fabric/platform/react/core/ConcreteComponentDescriptor.h
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// RN has some weird include directory redirections..this forwards the include to the right place
-#include <react/renderer/core/ConcreteComponentDescriptor.h>
